services/gateway/requirements.txt
Django==3.2.7
pytz==2021.1
sqlparse==0.4.2
djangorestframework==3.12.4
requests==2.25.1
